Update segcache to upstream with CAS generation-token fix
Upstream cache-rs fixed the CAS false-success regression by encoding a
per-segment generation into a 64-bit token (pelikan-io/cache-rs#24).
Adopt the new commit:

- Drop the u32 narrowing casts in entrystore now that the storage cas()
  method takes the full u64 token from the memcache protocol.
- Rework the 'cas stored' integration test to fetch the real cas token
  via gets instead of assuming tokens are a counter starting at 1; the
  token is opaque to clients under the new location+generation scheme.

+++ src/entrystore/src/segcache/memcache.rs
@@ -325,7 +325,7 @@ impl Storage for Seg {
                     v,
                     Some(&cas.flags().to_be_bytes()),
                     ttl,
-                    cas.cas() as u32,
+                    cas.cas(),
                 ) {
                     Ok(_) => Response::stored(cas.noreply()),
                     Err(SegcacheError::NotFound) => Response::not_found(cas.noreply()),
@@ -338,7 +338,7 @@ impl Storage for Seg {
                     cas.value(),
                     Some(&cas.flags().to_be_bytes()),
                     ttl,
-                    cas.cas() as u32,
+                    cas.cas(),
                 ) {
                     Ok(_) => Response::stored(cas.noreply()),
                     Err(SegcacheError::NotFound) => Response::not_found(cas.noreply()),
@@ -352,7 +352,7 @@ impl Storage for Seg {
                 cas.value(),
                 Some(&cas.flags().to_be_bytes()),
                 ttl,
-                cas.cas() as u32,
+                cas.cas(),
             ) {
                 Ok(_) => Response::stored(cas.noreply()),
                 Err(SegcacheError::NotFound) => Response::not_found(cas.noreply()),
